separate participation tokens to separate tab
